+++ chain/tests/test_coinbase_maturity.rs
@@ -24,13 +24,13 @@ use std::fs;
 use std::sync::Arc;
 
 use chain::types::*;
-use wallet::libtransaction::build;
+use core::consensus;
+use core::core::OutputIdentifier;
 use core::core::target::Difficulty;
 use core::core::transaction;
-use core::core::OutputIdentifier;
-use core::consensus;
 use core::global;
 use core::global::ChainTypes;
+use wallet::libtransaction::build;
 
 use keychain::Keychain;
 use wallet::libtransaction;
