+++ core/tests/block.rs
@@ -19,17 +19,17 @@ extern crate grin_wallet as wallet;
 
 pub mod common;
 
-use grin_core::core::{Block, BlockHeader, CompactBlock, KernelFeatures, OutputFeatures};
-use grin_core::core::hash::Hashed;
-use grin_core::core::block::Error;
-use grin_core::core::id::{ShortId, ShortIdentifiable};
-use wallet::libtransaction::build::{self, input, output, with_fee};
 use common::{new_block, tx1i2o, tx2i1o, txspend1i1o};
-use keychain::Keychain;
 use grin_core::consensus::{BLOCK_OUTPUT_WEIGHT, MAX_BLOCK_WEIGHT};
-use grin_core::ser;
+use grin_core::core::block::Error;
+use grin_core::core::hash::Hashed;
+use grin_core::core::id::{ShortId, ShortIdentifiable};
+use grin_core::core::{Block, BlockHeader, CompactBlock, KernelFeatures, OutputFeatures};
 use grin_core::global;
+use grin_core::ser;
+use keychain::Keychain;
 use std::time::Instant;
+use wallet::libtransaction::build::{self, input, output, with_fee};
 
 use util::{secp, secp_static};
