+++ wallet/src/libtransaction/mod.rs
@@ -26,4 +26,19 @@ pub mod build;
 pub mod error;
 pub mod proof;
 pub mod reward;
-pub mod transaction;
+pub mod slate;
+
+use core::consensus;
+use core::core::Transaction;
+
+const DEFAULT_BASE_FEE: u64 = consensus::MILLI_GRIN;
+
+/// Transaction fee calculation
+pub fn tx_fee(input_len: usize, output_len: usize, proof_len: usize, base_fee: Option<u64>) -> u64 {
+	let use_base_fee = match base_fee {
+		Some(bf) => bf,
+		None => DEFAULT_BASE_FEE,
+	};
+
+	(Transaction::weight(input_len, output_len, proof_len) as u64) * use_base_fee
+}
