+++ wallet/src/libtransaction/slate.rs
@@ -17,17 +17,15 @@
 use rand::thread_rng;
 use uuid::Uuid;
 
-use core::consensus;
 use core::core::{amount_to_hr_string, Committed, Transaction};
 use keychain::{BlindSum, BlindingFactor, Keychain};
 use libtransaction::error::Error;
-use libtransaction::{aggsig, build};
+use libtransaction::{aggsig, build, tx_fee};
 
 use util::secp::Signature;
 use util::secp::key::{PublicKey, SecretKey};
 use util::{secp, LOGGER};
 
-const DEFAULT_BASE_FEE: u64 = consensus::MILLI_GRIN;
 
 /// Public data for each participant in the slate
 
@@ -381,12 +379,4 @@ impl Slate {
 	}
 }
 
-/// Transaction fee calculation
-pub fn tx_fee(input_len: usize, output_len: usize, proof_len: usize, base_fee: Option<u64>) -> u64 {
-	let use_base_fee = match base_fee {
-		Some(bf) => bf,
-		None => DEFAULT_BASE_FEE,
-	};
 
-	(Transaction::weight(input_len, output_len, proof_len) as u64) * use_base_fee
-}
