+++ wallet/src/sender.rs
@@ -18,9 +18,9 @@ use client;
 use core::core::amount_to_hr_string;
 use core::ser;
 use failure::ResultExt;
-use libwallet::selection;
 use keychain::{Identifier, Keychain};
-use libtransaction::{build, transaction};
+use libtransaction::{build, tx_fee};
+use libwallet::selection;
 use receiver::TxWrapper;
 use types::*;
 use util;
@@ -163,7 +163,7 @@ pub fn issue_burn_tx(
 
 	debug!(LOGGER, "selected some coins - {}", coins.len());
 
-	let fee = transaction::tx_fee(coins.len(), 2, selection::coins_proof_count(&coins), None);
+	let fee = tx_fee(coins.len(), 2, selection::coins_proof_count(&coins), None);
 	let (mut parts, _) =
 		selection::inputs_and_change(&coins, config, keychain, current_height, amount, fee)?;
